@@ -1,20 +1,62 @@
-// import 'package:bengkol/login.dart';
-// import 'package:bengkol/login.dart';
-// import 'package:bengkol/sign_in.dart';
-// import 'package:firebase_auth/firebase_auth.dart';
-// import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:bengkol/sign_in1.dart';
+import 'package:flutter_application_6/login_menu.dart';
+import 'package:flutter_application_6/page_splash.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(home: SignInPage());
+    return const MaterialApp(
+      home: SplashScr(),
+    );
+  }
+}
+
+// splash screen
+class SplashScr extends StatelessWidget {
+  const SplashScr({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+        // title: 'bengkol',
+        home: PageSplash());
+  }
+}
+
+class MyPage extends StatefulWidget {
+  const MyPage({super.key});
+
+  @override
+  State<MyPage> createState() => _MyPageState();
+}
+
+class _MyPageState extends State<MyPage> {
+  Future<FirebaseApp> _initializeFirebase() async {
+    FirebaseApp firebaseApp = await Firebase.initializeApp();
+    return firebaseApp;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: FutureBuilder(
+        future: _initializeFirebase(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.done) {
+            return const LoginMenu();
+          }
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        },
+      ),
+    );
   }
 }
